refactor: post layout and page design, validation function
