+++ app/global.css
@@ -81,7 +81,7 @@ h4 {
 
 pre {
   background-color: #16161e;
-  border-radius: 0.5rem;
+  border-radius: 0.3rem;
   overflow-x: auto;
   padding: 1rem;
   margin: 1.5rem 0;
