+++ app/post/[slug]/page.tsx
@@ -1,7 +1,6 @@
 import { PostHeader } from "@/app/post/[slug]/post-header";
 import { notFound } from "next/navigation";
 import { getPostSlugs } from "@/lib/post";
-import { Fragment } from "react";
 
 import path from "node:path";
 
@@ -17,10 +16,12 @@ export default async function PostPage({ params }: PostPageProps) {
       `@/content/post/${slug}.mdx`
     );
     return (
-      <Fragment>
+      <div className="flex flex-col gap-2 mt-16">
         <PostHeader metadata={metadata} />
-        <Post />
-      </Fragment>
+        <div data-slot="post-content">
+          <Post />
+        </div>
+      </div>
     );
   } catch (error) {
     console.error(error);
