+++ app/post/page.tsx
@@ -3,10 +3,10 @@ import type { SearchParams } from "nuqs/server";
 
 import { loadPaginationParams } from "@/components/search-params";
 import { ArrowLeftIcon, ArrowRightIcon } from "lucide-react";
+import { validateMetadata, getPostSlugs } from "@/lib/post";
 import { Separator } from "@/components/ui/separator";
 import { Button } from "@/components/ui/button";
 import { PostCard } from "@/app/post/postcard";
-import { getPostSlugs } from "@/lib/post";
 import { FilePlus } from "lucide-react";
 
 import {
@@ -20,6 +20,13 @@ import path from "node:path";
 type PageProps = {
   searchParams: Promise<SearchParams>;
 };
+
+const requiredMetadata: string[] = [
+  "title",
+  "summary",
+  "publishedAt",
+  "readTime",
+];
 
 export default async function PostsPage({ searchParams }: PageProps) {
   const posts: Post[] = await getPost();
@@ -132,9 +139,11 @@ async function getPost() {
   const posts = await Promise.all(
     postsSlugs.map(async (slug: string) => {
       try {
-        const { metadata } = await import(`@/content/post/${slug}.mdx`);
+        const { metadata } = (await import(`@/content/post/${slug}.mdx`)) as {
+          metadata: PostMetadata;
+        };
 
-        if (!validatePostMetadata(metadata, slug)) return null;
+        if (!validateMetadata(metadata, requiredMetadata, slug)) return null;
 
         return { ...metadata, slug } as Post;
       } catch (error) {
@@ -145,17 +154,4 @@ async function getPost() {
   );
 
   return posts.filter(Boolean) as Post[];
-}
-
-function validatePostMetadata(
-  metadata: PostMetadata,
-  slug: string
-): metadata is Post {
-  const required = ["title", "summary", "publishedAt", "readTime"];
-  const missing = required.filter((field) => !(field in metadata));
-  if (missing.length > 0) {
-    console.warn(`Post ${slug} missing required fields:`, missing);
-    return false;
-  }
-  return true;
 }
